LPD-77908 [FE] Playwright: Moving and fixing related existing tests to CMP module
